fix apiUrl
