+++ eventually/static/src/containers/fileUpload/FileUploadService.js
@@ -1,6 +1,7 @@
+import {apiUrl} from './helper';
 import axios from 'axios';
 
-const request_path = '/api/v1/img/handle/';
+const request_path = apiUrl + 'img/handle/';
 const request_timeout = 30000;
 
 const sendFile = data => {
