+++ eventually/static/src/containers/profile/ProfileView.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import {getImageUrl} from '../../helper';
-import {getUserId} from '../../helper/utils';
+import {getImageUrl} from './helper';
+import {getUserId} from './helper';
 import ProfileEdit from './ProfileEdit';
 import {getProfileService} from './ProfileService';
 import RaisedButton from 'material-ui/RaisedButton';
