+++ eventually/static/src/containers/teamList/EditTeamDialog.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Dialog from 'material-ui/Dialog';
 import FlatButton from 'material-ui/FlatButton';
 import TextField from 'material-ui/TextField';
-import FileUpload from '../fileUpload/FileUpload';
+import FileUpload from './containers';
 import {teamServicePut} from './teamService';
 
 export default class EditTeamDialog extends React.Component {
