+++ eventually/static/src/containers/topicList/TopicItem.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Card, CardActions, CardHeader, CardText } from 'material-ui/Card';
 import RaisedButton from 'material-ui/RaisedButton';
-import TopicDialog from '../topicList/TopicDialog';
 import {lightGreen400} from 'material-ui/styles/colors';
+import {TopicDialog} from './containers';
 
 const cardTextstyle = {
     color: '#455A64',
