--- eventually/static/src/helper/utils.js
@@ -1,13 +1,0 @@
-export const isLogged = () => document.cookie.indexOf('sessionid') !== -1;
-
-export const getUserId = () => {
-    let cookies = document.cookie.split(';').map(cookie => cookie.replace(' ', ''));
-    const name = 'user_id=';
-
-    for (let i=0; i < cookies.length; i++) {
-        if(cookies[i].indexOf(name) !== -1) {
-            return +cookies[i].substring(name.length, cookies[i].length);
-        }
-    }
-    return null;
-};
