+++ eventually/static/src/layout.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Paper from 'material-ui/Paper';
 import MainRouter from './mainRouter';
-import {isLogged} from './helper/utils';
+import {isLogged} from './helper';
 import SignRouter from './signRouter';
 import Header from './components/header/Header';
 import Sign from './containers/registerLogin/Sign';
